requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
python-dotenv==1.0.1
gspread==6.1.4
google-auth==2.36.0
google-auth-oauthlib==1.2.1

